# backend/swagger.yaml
openapi: 3.0.0
info:
  title: NUSeek API
  version: 1.0.0
  description: API for connecting help seekers and helpers
servers:
  - url: http://localhost:3000
  - url: https://your-deployed-url.onrender.com
...
paths:
...
  /chats/rooms/{roomId}:
    get:
      tags:
        - Chat
      summary: Fetch all messages in a chat room
      parameters:
        - in: path
          name: roomId
          required: true
          schema:
            type: integer
      responses:
        200:
          description: List of messages
        500:
          description: Server error
    post:
      tags:
        - Chat
      summary: Create a new message in a chat room
      parameters:
        - in: path
          name: roomId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sender_id, receiver_id, content]
              properties:
                sender_id:
                  type: integer
                receiver_id:
                  type: integer
                content:
                  type: string
      responses:
        201:
          description: Message created successfully
        400:
          description: Missing required fields
        500:
          description: Server error
  /chats/rooms/users/{userId}:
    get:
      tags:
        - Chat
      summary: Fetch all chat rooms for a user
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
      responses:
        200:
          description: List of rooms with last message
        500:
          description: Server error
  /chats/messages/{messageId}:
    put:
      tags:
        - Chat
      summary: Update message read status
      parameters:
        - in: path
          name: messageId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [isRead]
              properties:
                isRead:
                  type: boolean
      responses:
        200:
          description: Message updated successfully
        404:
          description: Message not found
        500:
          description: Server error

  # TASKS ROUTES
...